# Conflicts:
#	src/test/java/cn/diffpi/resource/Text.java

+++ src/main/java/cn/diffpi/resource/file/FileResource.java
@@ -2,24 +2,16 @@ package cn.diffpi.resource.file;
 
 import cn.diffpi.kit.StrKit;
 import cn.diffpi.kit.video.FileUtil;
-import cn.diffpi.kit.video.ScanFileUtil;
 import cn.diffpi.resource.ApiResource;
 import cn.diffpi.resource.module.video.model.HiVideo;
 import cn.diffpi.resource.platform.config.model.HiConfig;
-import cn.dreampie.common.http.HttpMessage;
 import cn.dreampie.common.http.exception.HttpException;
 import cn.dreampie.common.http.result.HttpStatus;
-import cn.dreampie.orm.transaction.Transaction;
 import cn.dreampie.route.annotation.API;
 import cn.dreampie.route.annotation.DELETE;
-import cn.dreampie.route.annotation.GET;
 import cn.dreampie.route.annotation.POST;
-import cn.dreampie.route.core.multipart.FILE;
 
 import java.io.File;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 /**
  * 文件请求类
@@ -118,16 +110,21 @@ public class FileResource extends ApiResource {
         if (StrKit.notBlank(ids)) {
             for (String id : ids.split(",")) {
                 HiVideo hiVideo = HiVideo.dao.findById(id);
+                if (hiVideo == null) {
+                    throw new HttpException(HttpStatus.NOT_FOUND, "paramErr", id + " 对应的视频不存在！");
+                }
                 String path = hiVideo.get("path");
                 if (StrKit.notBlank(path)) {
                     // 获取配置的删除目录路径
                     String deletePath = HiConfig.dao.getConfigValue("deletePath");
                     if (StrKit.isBlank(deletePath)) {
-                        throw new HttpException(HttpStatus.NOT_FOUND, "尚未配置回收站！");
+                        throw new HttpException(HttpStatus.NOT_FOUND,"paramErr", "尚未配置回收站！");
                     }
                     // 移动成功就删除记录
                     if (FileUtil.moveFile(path, deletePath)) {
                         hiVideo.delete();
+                    } else {
+                        throw new HttpException(HttpStatus.NOT_FOUND,"paramErr", "移动失败！");
                     }
                 }
             }
